backend/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
